useCanEditDetectors and always show header checkbox
