+++ static/app/views/detectors/components/detectorListTable/detectorListRow.tsx
@@ -1,7 +1,6 @@
 import styled from '@emotion/styled';
 
 import {Checkbox} from 'sentry/components/core/checkbox';
-import {Flex} from 'sentry/components/core/layout';
 import Placeholder from 'sentry/components/placeholder';
 import {SimpleTable} from 'sentry/components/tables/simpleTable';
 import {IssueCell} from 'sentry/components/workflowEngine/gridCell/issueCell';
@@ -23,18 +22,17 @@ export function DetectorListRow({detector, selected, onSelect}: DetectorListRowP
       variant={detector.enabled ? 'default' : 'faded'}
       data-test-id="detector-list-row"
     >
-      <SimpleTable.RowCell>
-        <Flex gap="md">
-          <CheckboxWrapper>
-            <Checkbox
-              checked={selected}
-              onChange={() => onSelect(detector.id)}
-              className="select-row"
-            />
-          </CheckboxWrapper>
-
-          <DetectorLink detector={detector} />
-        </Flex>
+      <StyledRowCell data-column-name="selected">
+        <CheckboxWrapper>
+          <Checkbox
+            checked={selected}
+            onChange={() => onSelect(detector.id)}
+            className="select-row"
+          />
+        </CheckboxWrapper>
+      </StyledRowCell>
+      <SimpleTable.RowCell data-column-name="name">
+        <DetectorLink detector={detector} />
       </SimpleTable.RowCell>
       <SimpleTable.RowCell data-column-name="type">
         <DetectorTypeCell type={detector.type} />
@@ -90,9 +88,10 @@ const DetectorSimpleTableRow = styled(SimpleTable.Row)`
 `;
 
 const CheckboxWrapper = styled('div')`
-  width: 20px;
-  height: 20px;
-  display: flex;
-  align-items: center;
-  flex-shrink: 0;
+  height: 100%;
+  align-self: flex-start;
+`;
+
+const StyledRowCell = styled(SimpleTable.RowCell)`
+  height: 100%;
 `;
