+++ static/app/views/detectors/utils/useCanEditDetector.tsx
@@ -1,7 +1,8 @@
 import {hasEveryAccess} from 'sentry/components/acl/access';
-import type {DetectorType} from 'sentry/types/workflowEngine/detectors';
+import type {Detector, DetectorType} from 'sentry/types/workflowEngine/detectors';
 import useOrganization from 'sentry/utils/useOrganization';
 import useProjectFromId from 'sentry/utils/useProjectFromId';
+import useProjects from 'sentry/utils/useProjects';
 import {detectorTypeIsUserCreateable} from 'sentry/views/detectors/utils/detectorTypeConfig';
 
 export function useCanEditDetector({
@@ -26,4 +27,49 @@ export function useCanEditDetector({
 
   // For non-user-createable detectors (such as error, n+1 etc.), only team admins can edit.
   return project.access.includes('alerts:write');
+}
+
+export function useCanEditDetectors({detectors}: {detectors: Detector[]}) {
+  const organization = useOrganization();
+  const {projects} = useProjects();
+
+  const projectToDetectors: Record<string, Detector[]> = {};
+
+  if (
+    organization.access.includes('org:write') ||
+    organization.access.includes('org:admin')
+  ) {
+    return true;
+  }
+
+  for (const detector of detectors) {
+    const projectId = detector.projectId;
+    if (!projectToDetectors[projectId]) {
+      projectToDetectors[projectId] = [];
+    }
+    projectToDetectors[projectId]?.push(detector);
+  }
+
+  for (const projectId of Object.keys(projectToDetectors)) {
+    const project = projects.find(p => p.id === projectId) ?? undefined;
+    if (!project) {
+      return false;
+    }
+
+    if (hasEveryAccess(['alerts:write', 'project:read'], {organization, project})) {
+      // team admins can modify all detectors for projects they have admin access to
+      if (project.access.includes('project:write')) {
+        continue;
+      }
+      // members can modify only user-createable detectors for projects they have access to
+      if (
+        projectToDetectors[projectId]?.every(d => detectorTypeIsUserCreateable(d.type))
+      ) {
+        continue;
+      }
+    }
+    return false;
+  }
+
+  return true;
 }
